requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
fpdf==1.7.2
lxml==6.0.2
packaging==25.0
pefile==2023.2.7
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-docx==1.2.0
pywin32-ctypes==0.2.3
typing_extensions==4.15.0
